requirements.txt
discord.py==1.7.3

dontenv==0.20.0